feat: 역할에 따라 컴포넌트 분리
