+++ src/desktop/Works/types/worksType.ts
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export interface HoveredStudioInfoProps {
+  description: string;
+}
+
+export interface HoveredStudioImgProps {
+  img: React.ReactNode;
+}
